requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.11.0
Django==3.1.12
django-cors-headers==3.8.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.6
drf-yasg==1.20.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.2
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.0.1
jsonschema==3.2.0
lru-dict==1.1.7
magic-admin==0.0.5
MarkupSafe==2.0.1
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
packaging==21.0
parsimonious==0.8.1
protobuf==3.17.3
psycopg2==2.9.1
pycryptodome==3.10.1
pymemcache==3.5.0
pymongo==3.12.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-memcached==1.59
pytz==2021.1
pywin32==301
requests==2.26.0
rlp==2.0.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.2.4
toolz==0.11.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
varint==1.0.2
web3==5.23.1
websockets==9.1
yarl==1.6.3
